# openapi.yaml
...
openapi: 3.1.0
info:
  title: TorrGate API
  version: 1.0.0
  description: High-performance TypeScript API gateway and proxy for torrent trackers, adhering to the canonical Jackett REST v2.0 path space.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v2.0/indexers
    description: Canonical Jackett REST v2.0 endpoint space
security:
  - ApiKeyAuth: []
  - ApiKeyQuery: []
  - BearerAuth: []
  - {}
paths:
...
  /{indexer}/details/{id}:
    get:
      summary: Get topic details by ID
      description: Retrieves detailed metadata for a specific topic/release by its tracker ID.
      operationId: getTopicDetails
      parameters:
        - name: indexer
          in: path
          required: true
          description: Tracker indexer name or identifier
          schema:
            type: string
            example: "rutracker-ru"
        - name: id
          in: path
          required: true
          description: Topic or torrent ID
          schema:
            type: string
            example: "100001"
      responses:
        "200":
          description: Topic details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicDetails"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  responses:
...
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Too many failed API key attempts from this client; retry after the one-minute window
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard API error structure
      required:
        - error
        - message
        - statusCode
        - success
      properties:
        error:
          type: string
          example: "NotFound"
        message:
          type: string
          example: "Indexer 'unknown' not found"
        statusCode:
          type: integer
          example: 404
        success:
          type: boolean
          example: false
...
    TopicDetails:
      type: object
      description: Detailed metadata for a torrent topic
      required:
        - actors
        - audioTranslation
        - category
        - description
        - director
        - duration
        - id
        - imdbUrl
        - infoHash
        - kinopoiskUrl
        - magnetUri
        - name
        - posterUrl
        - releaseCountry
        - torrentUrl
        - url
        - year
      properties:
        actors:
          type: array
          items:
            type: string
          example: ["Актер Первый", "Актер Второй"]
        audioTranslation:
          type: string
          example: "Профессиональный"
        category:
          type: string
          example: "Фильмы"
        description:
          type: string
          example: "Вымышленное описание для примера."
        director:
          type: string
          example: "Тестовый Режиссер"
        duration:
          type: string
          example: "01:45:00"
        files:
          type: array
          items:
            $ref: "#/components/schemas/TorrentFile"
        id:
          type: string
          example: "100001"
        imdbId:
          type: string
          example: "1234567"
        imdbUrl:
          type: string
          example: "https://www.imdb.com/title/tt1234567/"
        infoHash:
          type: string
          example: "0123456789abcdef0123456789abcdef01234567"
        kinopoiskId:
          type: string
          example: "7654321"
        kinopoiskUrl:
          type: string
          example: "https://www.kinopoisk.ru/film/7654321/"
        magnetUri:
          type: string
          example: "magnet:?xt=urn:btih:0123456789abcdef0123456789abcdef01234567"
        name:
          type: string
          example: "Пример Фильма / Example Movie (2026) BDRip 1080p"
        posterUrl:
          type: string
          example: "https://example.com/poster.jpg"
        releaseCountry:
          type: string
          example: "Примерная Страна"
        torrentUrl:
          type: string
          example: "https://rutracker.org/forum/dl.php?t=100001"
        url:
          type: string
          example: "https://rutracker.org/forum/viewtopic.php?t=100001"
        year:
          type: string
          example: "2026"
    TorrentFile:
      type: object
      description: Individual file within a torrent release
      required:
        - name
        - size
      properties:
        name:
          type: string
          example: "sample.movie.2026.1080p.mkv"
        size:
          type: string
          example: "1.45 GB"
        sizeBytes:
          type: integer
          example: 1556925645
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Optional API key passed via HTTP header
    ApiKeyQuery:
      type: apiKey
      in: query
      name: apikey
      description: Optional API key passed via query parameter
    BearerAuth:
      type: http
      scheme: bearer
      description: Optional API key passed as a Bearer token